Refactoring ActivityPanel
