--- frog/imports/ui/GraphEditor/SidePanel/ActivityPanel.js
@@ -1,218 +1,0 @@
-// @flow
-import React, { Component } from 'react';
-import { createContainer } from 'meteor/react-meteor-data';
-import { FormGroup, ControlLabel, FormControl } from 'react-bootstrap';
-import {
-  EnhancedForm,
-  ChangeableText,
-  type ActivityPackageT
-} from 'frog-utils';
-import FlexView from 'react-flexview';
-
-import { Activities, addActivity } from '/imports/api/activities';
-import { activityTypes, activityTypesObj } from '/imports/activityTypes';
-import Preview from '../Preview';
-import { RenameField } from '../Rename';
-import { connect } from '../store';
-import FileForm from './fileUploader';
-import ListComponent from './ListComponent';
-
-const SelectAttribute = connect(({ activity, onChange, store: { valid } }) =>
-  <FormGroup controlId="selectGrouping">
-    <FlexView>
-      <FlexView vAlignContent="center" marginRight="10px">
-        <ControlLabel>Group by attribute</ControlLabel>
-      </FlexView>
-      <FlexView vAlignContent="center">
-        {valid && valid.social[activity._id]
-          ? <FormControl
-              onChange={e => onChange(e.target.value)}
-              componentClass="select"
-              defaultValue={activity.groupingKey}
-            >
-              {[undefined, ...valid.social[activity._id]].map(x =>
-                <option value={x} key={x}>
-                  {x}
-                </option>
-              )}
-            </FormControl>
-          : <span style={{ color: 'red' }}>
-              No attributes provided, add social operator
-            </span>}
-        {valid.social[activity._id].includes(activity.groupingKey)
-          ? null
-          : <p style={{ color: 'red' }}>
-              You previously selected the {activity.groupingKey} attribute, but
-              no social operator is currently providing that attribute, please
-              reconfigure the incoming social operators, or select another
-              attribute
-            </p>}
-      </FlexView>
-    </FlexView>
-  </FormGroup>
-);
-
-class ChooseActivityTypeComp extends Component {
-  state: { expanded: ?string, searchStr: string, showInfo: ?string };
-  inputRef: any;
-
-  constructor(props) {
-    super(props);
-    this.state = { expanded: null, searchStr: '', showInfo: null };
-    this.inputRef = null;
-  }
-
-  componentDidMount() {
-    this.inputRef.focus();
-  }
-
-  render() {
-    const select = activityType => {
-      Activities.update(this.props.activity._id, {
-        $set: { activityType: activityType.id }
-      });
-      this.props.store.addHistory();
-    };
-
-    const changeSearch = e =>
-      this.setState({
-        expanded: null,
-        searchStr: e.target.value.toLowerCase()
-      });
-
-    const filteredList = activityTypes.filter(
-      x =>
-        x.meta.name.toLowerCase().includes(this.state.searchStr) ||
-        x.meta.shortDesc.toLowerCase().includes(this.state.searchStr) ||
-        x.meta.description.toLowerCase().includes(this.state.searchStr)
-    );
-
-    return (
-      <div style={{ height: '100%' }}>
-        <div style={{ display: 'flex', flexDirection: 'row' }}>
-          <h4>Please select activity type</h4>
-          <div
-            className="input-group"
-            style={{ top: '5px', left: '10px', width: '250px' }}
-          >
-            <span className="input-group-addon" id="basic-addon1">
-              <span className="glyphicon glyphicon-search" aria-hidden="true" />
-            </span>
-            <input
-              ref={ref => (this.inputRef = ref)}
-              type="text"
-              style={{ zIndex: 0 }}
-              onChange={changeSearch}
-              className="form-control"
-              placeholder="Search for..."
-              aria-describedby="basic-addon1"
-            />
-          </div>
-        </div>
-        <div
-          className="list-group"
-          style={{ height: '730px', width: '100%', overflow: 'scroll' }}
-        >
-          {filteredList.length === 0
-            ? <div
-                style={{
-                  marginTop: '20px',
-                  marginLeft: '10px',
-                  fontSize: '40px'
-                }}
-              >
-                No result
-              </div>
-            : filteredList.map((x: ActivityPackageT) =>
-                <ListComponent
-                  hasPreview={x.meta.exampleData !== undefined}
-                  onSelect={() => select(x)}
-                  showExpanded={this.state.expanded === x.id}
-                  expand={() => this.setState({ expanded: x.id })}
-                  key={x.id}
-                  onPreview={() => this.setState({ showInfo: x.id })}
-                  object={x}
-                  searchS={this.state.searchStr}
-                  eventKey={x.id}
-                />
-              )}
-        </div>
-        {this.state.showInfo !== null &&
-          <Preview
-            activityTypeId={this.state.showInfo}
-            dismiss={() => this.setState({ showInfo: null })}
-          />}
-      </div>
-    );
-  }
-}
-
-const EditClass = props => {
-  const activity = props.activity;
-  const graphActivity = props.store.activityStore.all.find(
-    act => act.id === activity._id
-  );
-
-  return (
-    <div>
-      <div style={{ backgroundColor: '#eee' }}>
-        <h3>
-          <ChangeableText
-            EditComponent={RenameField}
-            activityId={activity._id}
-            value={graphActivity.title}
-            onChange={grp =>
-              addActivity(activity.activityType, null, activity._id, grp)}
-          />
-        </h3>
-        <font size={-3}>
-          <i>
-            {`Type: ${activityTypesObj[activity.activityType].meta.name}
-                     (${activity.activityType})`}
-            <br />
-            {`Starting after ${graphActivity.startTime} min., running for ${graphActivity.length} min.`}
-          </i>
-        </font>
-        {activity.plane === 2 &&
-          <SelectAttribute
-            activity={activity}
-            onChange={grp =>
-              addActivity(activity.activityType, null, activity._id, grp)}
-          />}
-      </div>
-      <EnhancedForm
-        schema={activityTypesObj[activity.activityType].config}
-        UISchema={activityTypesObj[activity.activityType].configUI}
-        onChange={data => {
-          addActivity(
-            activity.activityType,
-            data.formData,
-            activity._id,
-            null,
-            data.errors.length > 0
-          );
-          props.store.refreshValidate();
-        }}
-        formData={activity.data}
-        liveValidate
-      >
-        <div />
-      </EnhancedForm>
-      <FileForm />
-    </div>
-  );
-};
-
-const EditActivity = connect(EditClass);
-const ChooseActivityType = connect(ChooseActivityTypeComp);
-
-export default createContainer(
-  ({ id }) => ({ activity: Activities.findOne(id) }),
-  ({ activity }) => {
-    if (activity.activityType && activity.activityType !== '') {
-      return <EditActivity activity={activity} />;
-    } else {
-      return <ChooseActivityType activity={activity} />;
-    }
-  }
-);
